requirements.txt
alembic==1.5.8
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.7.0
Flask-SQLAlchemy==2.5.1
googletrans==3.0.0
greenlet==1.0.0
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
Mako==1.1.4
MarkupSafe==1.1.1
nltk==3.5
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.3
psycopg2-binary==2.8.6
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.3.0
textblob==0.15.3
tqdm==4.59.0
tweepy==3.10.0
urllib3==1.26.4
Werkzeug==1.0.1
